refactor(core): rename display-context to display-provider, restructure use-entity-name
- providers/display-context -> providers/display-provider, matching the
  xxx-provider naming convention every other provider in core/providers
  follows (RepeatedLayoutProvider is the closest precedent: a nested,
  non-app-boundary context, not a single top-level DI wire-up, and it
  still lives under providers/ with that exact file/name split — raw
  context in <name>-context.ts, provider component in <name>-provider.tsx,
  hook in use-<name>-context.ts).
- Component renamed DisplayContext -> DisplayProvider to match; the raw
  React context itself is now named DisplayContext, in display-context.ts.
- use-entity-name.ts reordered newspaper-style: useEntityName (the hook
  most call sites reach for) leads, formatEntityName (the non-hook escape
  hatch for list rendering) follows, and the private toNavCase transform
  is last.
