+++ javascript/packages/core/components/breadcrumb-bar/phase-entity-list.tsx
@@ -2,7 +2,7 @@ import { useLocation, useNavigate } from 'react-router-dom-v5-compat';
 import { useStyletron } from 'baseui';
 
 import { formatEntityName } from '#core/hooks/use-entity-name/use-entity-name';
-import { useDisplayContext } from '#core/providers/display-context/use-display-context';
+import { useDisplayContext } from '#core/providers/display-provider/use-display-context';
 import { EntityItem } from './styled-components';
 
 import type { PhaseConfig } from '#core/types/common/studio-types';
