+++ javascript/packages/core/primitives.tsx
@@ -20,9 +20,9 @@ export { UserProvider } from '#core/providers/user-provider/user-provider';
 export { CellProvider } from '#core/providers/cell-provider/cell-provider';
 export { useCellProvider } from '#core/providers/cell-provider/use-cell-provider';
 export type { CellContextType } from '#core/providers/cell-provider/types';
-export { DisplayContext } from '#core/providers/display-context/display-context';
-export { useDisplayContext } from '#core/providers/display-context/use-display-context';
-export type { DisplayContextType } from '#core/providers/display-context/types';
+export { DisplayProvider } from '#core/providers/display-provider/display-provider';
+export { useDisplayContext } from '#core/providers/display-provider/use-display-context';
+export type { DisplayContextType } from '#core/providers/display-provider/types';
 
 // Hooks
 export { formatEntityName, useEntityName } from '#core/hooks/use-entity-name/use-entity-name';
