+++ javascript/packages/core/providers/display-provider/display-context.ts
@@ -6,4 +6,4 @@ import type { DisplayContextType } from './types';
  * `undefined` means no region has declared a display context — consumers
  * (e.g. `useEntityName`) treat that as "no casing transform".
  */
-export const DisplayContextValue = createContext<DisplayContextType | undefined>(undefined);
+export const DisplayContext = createContext<DisplayContextType | undefined>(undefined);
