+++ javascript/packages/core/providers/display-provider/use-display-context.ts
@@ -1,0 +1,11 @@
+import { useContext } from 'react';
+
+import { DisplayContext } from './display-context';
+
+/**
+ * Reads the nearest `DisplayProvider` ancestor's type, or `undefined` if
+ * nothing in the tree has declared one.
+ */
+export function useDisplayContext() {
+  return useContext(DisplayContext);
+}
