+++ javascript/packages/core/types/common/studio-types.ts
@@ -66,7 +66,7 @@ export interface PhaseEntityConfig<T extends object = object> {
   /**
    * Name of the entity as it appears within MA Studio. Should be plural, lower case
    * version of the name — casing is applied at render time based on the surrounding
-   * `DisplayContext` (see `useEntityName` in `#core/hooks/use-entity-name/use-entity-name`),
+   * `DisplayProvider` (see `useEntityName` in `#core/hooks/use-entity-name/use-entity-name`),
    * so config authors should supply the plain lower case form and not pre-case it.
    * An acronym that's part of the name (e.g. "AI") should still be written capitalized —
    * `useEntityName` only lowercases nothing and only capitalizes the first letter of
